+++ CHANGELOG.md
@@ -5,9 +5,10 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
-## 3.0.0-alpha.25 (2024-03-29)
 
-* (theimo1221) Re-add iRestSettings
+## **WORK IN PROGRESS**
+
+* (theimo1221) Reduce log spam
 
 ## 3.0.0